working with home bloc instead of search bloc
