+++ lib/home/bloc/home_bloc.dart
@@ -1,10 +1,12 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:in_app_update/in_app_update.dart';
 import 'package:injectable/injectable.dart';
 
-import '../../common/utils/env/flavor_config.dart';
+import '../../data/models/book.dart';
+import '../../data/models/songext.dart';
+import '../../data/repository/database_repository.dart';
+import '../../di/injectable.dart';
 
 part 'home_event.dart';
 part 'home_state.dart';
@@ -14,52 +16,42 @@ part 'home_bloc.freezed.dart';
 @injectable
 class HomeBloc extends Bloc<HomeEvent, HomeState> {
   HomeBloc() : super(const HomeState()) {
-    on<HomeCheckUpdates>(_onCheckUpdates);
-    on<HomeUpdateApp>(_onUpdateApp);
+    on<HomeFetchData>(_onFetch);
+    on<HomeSelect>(_onSelect);
+    on<HomeSearch>(_onSearch);
   }
 
-  Future<void> _onCheckUpdates(
-    HomeCheckUpdates event,
+  final _dbRepo = getIt<DatabaseRepository>();
+
+  Future<void> _onFetch(
+    HomeFetchData event,
     Emitter<HomeState> emit,
   ) async {
     emit(state.copyWith(status: Status.inProgress));
-    if (FlavorConfig.isProd()) {
-      try {
-        await InAppUpdate.checkForUpdate().then((updateInfo) {
-          if (updateInfo.updateAvailability ==
-              UpdateAvailability.updateAvailable) {
-            emit(
-              state.copyWith(
-                status: Status.updateFound,
-                feedback: 'There is a new app update',
-              ),
-            );
-          } else {
-            emit(state.copyWith(status: Status.success, feedback: ''));
-          }
-        });
-      } catch (e) {
-        emit(state.copyWith(status: Status.success, feedback: ''));
-      }
-    } else {
-      emit(state.copyWith(status: Status.success, feedback: ''));
-    }
+    var books = await _dbRepo.fetchBooks();
+    var songs = await _dbRepo.fetchSongExts();
+
+    emit(state.copyWith(status: Status.loaded, books: books, songs: songs));
   }
 
-  Future<void> _onUpdateApp(
-    HomeUpdateApp event,
+  Future<void> _onSelect(
+    HomeSelect event,
     Emitter<HomeState> emit,
   ) async {
-    InAppUpdate.performImmediateUpdate().then((appUpdateResult) {
-      if (appUpdateResult == AppUpdateResult.success) {
-        emit(
-          state.copyWith(
-            status: Status.updated,
-            feedback: 'Songlib successfully Updated',
-          ),
-        );
+    emit(state.copyWith(status: Status.inProgress));
+
+    List<SongExt> tempList = [];
+    for (int i = 0; i < state.songs.length; i++) {
+      if (state.songs[i].book == event.book.bookNo) {
+        tempList.add(state.songs[i]);
       }
-    });
+    }
+
+    emit(state.copyWith(status: Status.selected, filtered: tempList));
   }
 
+  Future<void> _onSearch(
+    HomeSearch event,
+    Emitter<HomeState> emit,
+  ) async {}
 }
