+++ lib/home/bloc/home_event.dart
@@ -2,7 +2,9 @@ part of 'home_bloc.dart';
 
 @freezed
 sealed class HomeEvent with _$HomeEvent {
-  const factory HomeEvent.checkUpdates() = HomeCheckUpdates;
+  const factory HomeEvent.fetch() = HomeFetchData;
 
-  const factory HomeEvent.updateApp() = HomeUpdateApp;
+  const factory HomeEvent.select(Book book) = HomeSelect;
+
+  const factory HomeEvent.search(String query) = HomeSearch;
 }
