+++ lib/navigator/route_names.dart
@@ -2,6 +2,7 @@ class RouteNames {
   RouteNames._();
 
   static const splash = 'splash';
+  static const appChecker = 'app_checker';
   static const init = 'init';
   static const login = 'login';
   static const signup = 'signup';
