+++ lib/user/app_checker/bloc/app_checker_event.dart
@@ -1,0 +1,8 @@
+part of 'app_checker_bloc.dart';
+
+@freezed
+sealed class AppCheckerEvent with _$AppCheckerEvent {
+  const factory AppCheckerEvent.checkUpdates() = AppCheckerCheckUpdates;
+
+  const factory AppCheckerEvent.updateApp() = AppCheckerUpdateApp;
+}
